backend/requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_migrate
flask-cors
flask-mail
pyclean
python-dotenv
pymysql
mysql-connector-python
email-validator